requirements.txt
selenium
webdriver-helper
pytest
pytest-rerunfailures
allure-pytest
openpyxl
pydantic